Crate ibc_core_host

Provides essential top-level traits designed for the seamless integration of host chains with ibc-rs. It streamlines access to the host’s storage, facilitating the efficient retrieval of states and metadata crucial for the execution of IBC logics.

Modules

types

Re-exports ICS-24 data structures from ibc-core-host-types crate.

Traits

ExecutionContext

Context to be implemented by the host that provides all “write-only” methods.

ValidationContext

Context to be implemented by the host that provides all “read-only” methods.

Type Aliases

ClientStateMut

Convenient type alias for ClientStateMut, providing access to client execution methods within the context.

ClientStateRef

Convenient type alias for ClientStateRef, providing access to client validation methods within the context.

ConsensusStateRef

Convenient type alias for ConsensusStateRef, providing access to client validation methods within the context.