ibc_core_host
Trait ExecutionContext

pub trait ExecutionContext: ValidationContext {
    type E: ClientExecutionContext;

    // Required methods
    fn get_client_execution_context(&mut self) -> &mut Self::E;
    fn increase_client_counter(&mut self) -> Result<(), HostError>;
    fn store_connection(
        &mut self,
        connection_path: &ConnectionPath,
        connection_end: ConnectionEnd,
    ) -> Result<(), HostError>;
    fn store_connection_to_client(
        &mut self,
        client_connection_path: &ClientConnectionPath,
        conn_id: ConnectionId,
    ) -> Result<(), HostError>;
    fn increase_connection_counter(&mut self) -> Result<(), HostError>;
    fn store_packet_commitment(
        &mut self,
        commitment_path: &CommitmentPath,
        commitment: PacketCommitment,
    ) -> Result<(), HostError>;
    fn delete_packet_commitment(
        &mut self,
        commitment_path: &CommitmentPath,
    ) -> Result<(), HostError>;
    fn store_packet_receipt(
        &mut self,
        receipt_path: &ReceiptPath,
        receipt: Receipt,
    ) -> Result<(), HostError>;
    fn store_packet_acknowledgement(
        &mut self,
        ack_path: &AckPath,
        ack_commitment: AcknowledgementCommitment,
    ) -> Result<(), HostError>;
    fn delete_packet_acknowledgement(
        &mut self,
        ack_path: &AckPath,
    ) -> Result<(), HostError>;
    fn store_channel(
        &mut self,
        channel_end_path: &ChannelEndPath,
        channel_end: ChannelEnd,
    ) -> Result<(), HostError>;
    fn store_next_sequence_send(
        &mut self,
        seq_send_path: &SeqSendPath,
        seq: Sequence,
    ) -> Result<(), HostError>;
    fn store_next_sequence_recv(
        &mut self,
        seq_recv_path: &SeqRecvPath,
        seq: Sequence,
    ) -> Result<(), HostError>;
    fn store_next_sequence_ack(
        &mut self,
        seq_ack_path: &SeqAckPath,
        seq: Sequence,
    ) -> Result<(), HostError>;
    fn increase_channel_counter(&mut self) -> Result<(), HostError>;
    fn emit_ibc_event(&mut self, event: IbcEvent) -> Result<(), HostError>;
    fn log_message(&mut self, message: String) -> Result<(), HostError>;

}

Context to be implemented by the host that provides all “write-only” methods.

Trait used for the top-level execute and dispatch entrypoints in the ibc-core crate.

Required Associated Types

type E: ClientExecutionContext

Required Methods

fn get_client_execution_context(&mut self) -> &mut Self::E

Retrieve the context that implements all clients’ ExecutionContext.

fn increase_client_counter(&mut self) -> Result<(), HostError>

Called upon client creation. Increases the counter, that keeps track of how many clients have been created.

fn store_connection( &mut self, connection_path: &ConnectionPath, connection_end: ConnectionEnd, ) -> Result<(), HostError>

Stores the given connection_end at path

fn store_connection_to_client( &mut self, client_connection_path: &ClientConnectionPath, conn_id: ConnectionId, ) -> Result<(), HostError>

Stores the given connection_id at a path associated with the client_id.

fn increase_connection_counter(&mut self) -> Result<(), HostError>

Called upon connection identifier creation (Init or Try process). Increases the counter which keeps track of how many connections have been created.

fn store_packet_commitment( &mut self, commitment_path: &CommitmentPath, commitment: PacketCommitment, ) -> Result<(), HostError>

Stores the given packet commitment at the given store path

fn delete_packet_commitment( &mut self, commitment_path: &CommitmentPath, ) -> Result<(), HostError>

Deletes the packet commitment at the given store path

fn store_packet_receipt( &mut self, receipt_path: &ReceiptPath, receipt: Receipt, ) -> Result<(), HostError>

Stores the given packet receipt at the given store path

fn store_packet_acknowledgement( &mut self, ack_path: &AckPath, ack_commitment: AcknowledgementCommitment, ) -> Result<(), HostError>

Stores the given packet acknowledgement at the given store path

fn delete_packet_acknowledgement( &mut self, ack_path: &AckPath, ) -> Result<(), HostError>

Deletes the packet acknowledgement at the given store path

fn store_channel( &mut self, channel_end_path: &ChannelEndPath, channel_end: ChannelEnd, ) -> Result<(), HostError>

Stores the given channel_end at a path associated with the port_id and channel_id.

fn store_next_sequence_send( &mut self, seq_send_path: &SeqSendPath, seq: Sequence, ) -> Result<(), HostError>

Stores the given nextSequenceSend number at the given store path

fn store_next_sequence_recv( &mut self, seq_recv_path: &SeqRecvPath, seq: Sequence, ) -> Result<(), HostError>

Stores the given nextSequenceRecv number at the given store path

fn store_next_sequence_ack( &mut self, seq_ack_path: &SeqAckPath, seq: Sequence, ) -> Result<(), HostError>

Stores the given nextSequenceAck number at the given store path

fn increase_channel_counter(&mut self) -> Result<(), HostError>

Called upon channel identifier creation (Init or Try message processing). Increases the counter, that keeps track of how many channels have been created.

fn emit_ibc_event(&mut self, event: IbcEvent) -> Result<(), HostError>

Emit the given IBC event

fn log_message(&mut self, message: String) -> Result<(), HostError>

Log the given message.